+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     - Create a single place where all redemptions can be redeemed
     - Cached provider, when available, is now always used when connecting to one's wallet
     - Nicer layout of the list of DAO cards on mobile devices
+    - more friendly display of error conditions (404 and uncaught exception errors)
   - Bugs Fixed
     - Align headers of table in proposal history
     - Re-add redeem for beneficiary button to proposal details page